feat: config to not fail loading module when shared lib fails to load
This commit makes it possible to specify a config option so that the
on_init callback in the sasl_auth module returns ok even when the shared
library don't load. This is useful, for example, for the EMQX message
broker as it makes it possible to start a system containing sasl_auth
even when some of its dependencies are missing on the system.